Add test for converting local time to UTC
