--- common/main/src/test/java/com/ivy/common/time/LocalUtcTimeConversionTest.kt
@@ -1,9 +1,0 @@
-package com.ivy.common.time
-
-import io.kotest.core.spec.style.StringSpec
-
-class LocalUtcTimeConversionTest : StringSpec({
-    "local <> utc" {
-        val local = timeNow()
-    }
-})
